soc: octeontx2: Fix npa_pf_func size in sso_hw_setconfig mbox msg
PF_FUNC should be 16bit and not 32bit.

Orabug: 30484033

+++ drivers/soc/marvell/octeontx2/mbox.h
@@ -707,7 +707,7 @@ struct ssow_lf_free_req {
 struct sso_hw_setconfig {
 	struct mbox_msghdr hdr;
 	u32	npa_aura_id;
-	u32	npa_pf_func;
+	u16	npa_pf_func;
 	u16	hwgrps;
 };
